Eliminate NaNs
Rarely, the BME280 can return NaN on sensor readings. Why exactly is
unclear, but what is certain is that NaN is not a valid JSON value. By
caching the last known-good reading for each sensor type, a reasonable
value can be reported even if temporarily none is available.

If a NaN would be returned during the very first reading, it is possible
that invalid JSON could still be emitted. This is extremely unlikely to
occur, however.